requirements.txt
psycopg2-binary==2.9.9
spotipy==2.23.0